Use bool array to track which mutants are enabled
Paving the way for having multiple mutants enabled, this change
introduces an array of booleans to track the mutations that are
enabled on a per file basis. At present, at most one element of this
array will be set to true. In a future change, environment variable
parsing will be updated to support multiple mutants.

Part of #55.