Allow mutation inside initializers
Refines the logic for deciding whether the mutation visitor is inside
a function so that expressions under variable declarations that are
themselves in functions are regarded as being in functions. This
allows mutating variable initializers inside functions, but still
disallows mutating the initializers of e.g. class members.

Care is also taken to avoid mutating constexpr initializers.

Fixes #50.