Only mutate l-values under implicit casts
Changes the way ++ and -- are applied during unary operator insertion,
so that they are only applied to l-values appearing directly under
expressions that produce r-values, and then in prefix form.

This allows the mutation to work both for C and C++, because even
though these prefix operators do not return l-values in C, if they are
only applied right before an implicitly cast to an r-value the results
they return will be acceptable.

For C++ it would be possible to apply these operators slightly more
generally, but it does not seem worth it, since mutating before each
use of an l-value in an expression should cover all the cases that
would be covered by e.g. mutating l-values as they are passed around
as parameters.

Fixes #88.