Restrict mutation of cast expressions
Restricts mutation of cast expressions to the case where the
expression inside the case is an l-value and the cast expression is an
r-value. This avoids a large number of cases where literal values are
implicitly case from signed to unsigned, or from double to float, and
get mutated twice as a result.

Fixes #121.