Do not remove self-inverse unary operators
Self-inverse unary operator removal is equivalent to insertion of
another instance of the unary operator, so is is redundant to perform
both mutations.

This change avoids the need for some checks to decide when operator
removal was redudant due to expressions having constant values. That
code had been identified to be buggy, which its removal also fixes.

Fixes #97.
Fixes #158.