Avoid lambdas where possible
When mutating C++ code, until now Dredd has used a lambda to capture
every argument to a replacement function. There have been two reasons
behind this. First, for short-circuiting operators (&& and ||) it is
important *not* to evaluate the second argument unless the first
argument evaluates to an appropriate value. Violating this could cause
the instrumented version of the original program to crash - e.g. an
expression such as:

y != 0 && ((x / y) > 1)

should only evaluate x / y when y is non-zero; turning && into a call
to a function that eagerly evaluates both the LHS and RHS would cause
a division by zero when y is zero.

The other reason is to more faithfully model source level mutations
when arguments have side effects. As an example, suppose we have an
expression:

a + (y = 2)

If we replace this with its LHS, we would like to mimic actually
replacing the text "a = (y = 2)" with the text "a", so that the side
effect does not happen. By representing arguments using lambdas we
allow this.

The problem with using lambdas everywhere by default is that generated
code size gets very large, which can lead to high compile times. There
is also associated runtime overhead when the compiled code is
executed.

This change to Dredd uses Clang's side effect analysis to ensure that
lambdas are only used either due operators being short-circuiting, or
due to potential side effects.

In relation to performance and code size, mutating the SPIRV-Tools
validator (all C++ files) and running all the validator unit tests, I
find:

Compilation time of SPIR-V validator + validator tests (averaged over
three runs)

- No Dredd instrumentation:                  32 seconds
- Dredd instrumentation, before this change: 75 seconds
- Dredd instrumentation, with this change:   46 seconds

Execution of validator tests (no mutations enabled, averaged over
three runs)

- No Dredd instrumentation:                   20 seconds
- Dredd instrumentation, before this change: 121 seconds
- Dredd instrumentation, with this change:    65 seconds

Combined size of validator test binaries:

- No Dredd instrumentation:                   132 Mb
- Dredd instrumentation, before this change: 1564 Mb
- Dredd instrumentation, with this change:    647 Mb