Work around structured binding issues
Due to what looks like a bug in the way source information is attached
to the Clang AST, expressions that appear in the context of structured
bindings have source ranges that end before they begin. This change
avoids mutating such expressions (because mutating them leads to
garbled source code text). It also adds a structured binding test
case, and has workaround to avoid problems with lambda capture and
structured bindings, support for which is not full-fledged in some
version of Clang.