Insert Dredd declarations at very start of file (#199)
Rather than looking for the first declaration in the file and
inserting Dredd declarations before that, declarations are now
inserted at the very beginning of the source file being mutated. This
means that they even precede preprocessor directives and comments. This
fixes an issue where declarations were being placed at invalid
positions in the presence of certain uses of preprocessor directives.

Fixes #198.