Fix braced initalization and binary zero optimizations
Previously, when mutating expressions such as `char test {24}`, the
braced initialization expression `{24}` would also get mutated (as noted
in #225). This lead to compilation errors. This is resolved by avoiding
the mutation of `InitListExpr`.

Also, the optimization checks for the case when both arguments to a
binary expression were zero was previously wrong (as noted in #222).
This is resolved.

Fixes: #222, #225.