Avoid mutation of array-size expression in C++ `new` expression
In C++, one can initialize a list of items of type `a` with the
expression `new a[2]{3, 4}`. Using a non-constant array size requires
the type `a` to have a zero-argument constructor. Since applying
mutation to the array size (i.e., `2` in the given example) would change
the constant array size to non-constant, and we aren't sure whether type
`a` contains a zero-argument constructor, we avoid mutating the array
size altogether.

Fixes #265 .