Rewrite sizes of certain constant-sized arrays
Avoids a problem where constant declarations that appear in the size
expression for a constant-sized array might have their initialisers
mutated, such that the size expression is no longer a compile time
constant. This leads to compilation errors in some scenarios
(including with Clang if the array is also initialised). The solution
is to rewrite the size expressions of such arrays, replacing the
original compile-time-constant expression with the value to which it
evaluates. The constants that make up the expression - which may also
be referred to elsewhere - can then have their initial values freely
mutated.

Fixes #263.