Rework protobuf representation of mutation tree
To avoid a restriction of the protobuf library that limits the
recursion depth of a protobuf message, this change reworks the way the
mutation tree is represented in protobuf format. The tree is now
organised as a flat list of nodes, and the children of a node are
identified via indices into this list.

Fixes #192.