Fix for issue with casts and initializer lists
Fixes a problem where an explicit cast was not being added to a
mutated expression in an initializer list where (after mutation) it is
required. The logic to handle this was present, but was designed to
expect the parents of the relevant AST node to have a specific
order. This change generalises the logic to allow for multiple
initializer list parents, because in the presence of templates a node
may have two initializer list parents: one for the uninstantated
version of the initializer list and the other for the instantiated
version.

Fixes #321.