Account for pasted macros when handling source ranges
Adjusts the logic for attempting to find the main file source range of
an AST node to take account of "pasted macros", which arise when the
preprocessor operator ## is used. In the presence of pasted macros,
the logic for determining whether an AST node corresponding to a macro
expansion has a meaningful source range in the main file was not
working correctly. This change detects pasted macros and
conservatively refuses to give a main file source range in their
presence.

Fixes #345.
Related issue: #338.